second commit

+++ OOP/src/Lab1/task1/Task1.java
@@ -4,7 +4,7 @@ public class Task1 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Enter your age:  ");
-        int age = scanner.nextInt();
+        byte age = scanner.nextByte();
         System.out.println(age);
         scanner.close();
     }
